Conflicts:
	build/sdk.atree
	ide/eclipse/.classpath
	tools/activitycreator/src/com/android/activitycreator/ActivityCreator.java
	tools/eclipse/scripts/create_adt_symlinks.sh
	tools/eclipse/scripts/create_all_symlinks.sh
	tools/eclipse/scripts/create_bridge_symlinks.sh
	tools/eclipse/scripts/create_common_symlinks.sh
	tools/eclipse/scripts/create_ddms_symlinks.sh
	tools/eclipse/scripts/create_editors_symlinks.sh
	tools/eclipse/scripts/create_test_symlinks.sh
	tools/scripts/build.template

+++ simulator/app/DeviceWindow.cpp
@@ -140,9 +140,9 @@ void DeviceWindow::ClampMouse(wxMouseEvent* pEvent)
 void DeviceWindow::OnUserEvent(UserEvent& event)
 {
     wxBitmap* pBitmap;
-    int displayIndex;
+    long displayIndex;
 
-    displayIndex = (int) event.GetData();
+    displayIndex = (long) event.GetData();
 
     //printf("GOT UAE %d\n", displayIndex);
